+++ taxi-frontend/src/app/gestor/pages/taxis-list/taxis-list.component.ts
@@ -2,7 +2,9 @@ import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 
 import { Taxi } from 'src/app/core/models/taxi';
+import { Turno } from 'src/app/core/models/turno';
 import { TaxiService } from 'src/app/core/services/taxi.service';
+import { TurnoService } from 'src/app/core/services/turno.service';
 
 @Component({
   selector: 'app-taxis-list',
@@ -12,10 +14,11 @@ import { TaxiService } from 'src/app/core/services/taxi.service';
 export class TaxisListComponent {
   taxis: Taxi[] = [];
   errorMessage: string = '';
-
+  turnos: Turno[] = [];
   
   constructor(
     private taxiService: TaxiService,
+    private turnoService: TurnoService,
     private router: Router
   ) {}
 
@@ -36,14 +39,37 @@ export class TaxisListComponent {
   }
 
   delete(taxi: Taxi): void {
-    this.taxiService.deleteTaxi(taxi._id).subscribe({
-      next: () => {
-        this.taxis = this.taxis.filter((h) => h !== taxi);
+    if (!taxi._id) {
+      this.errorMessage = 'ID do taxi inválido.';
+      return;
+    }
+
+    this.errorMessage = '';
+
+    this.turnoService.getAllTaxiShifts(taxi._id).subscribe({
+      next: (turnos) => {
+        const now = new Date();
+        this.turnos = turnos.filter(turno => new Date(turno.end) >= now);
+
+        if (this.turnos.length > 0) {
+          this.errorMessage = 'Taxi ainda tem turnos.';
+          return;
+        }
+
+        this.taxiService.deleteTaxi(taxi._id).subscribe({
+          next: () => {
+            this.taxis = this.taxis.filter((h) => h !== taxi);
+          },
+          error: (err) => {
+            console.error('Failed to delete taxi:', err);
+            this.errorMessage = 'Failed to delete taxi.';
+          },
+        });
       },
       error: (err) => {
-        console.error('Failed to delete taxi:', err);
-        this.errorMessage = 'Failed to delete taxi.';
-      },
+        console.error('Erro ao buscar turnos:', err);
+        this.errorMessage = 'Erro ao verificar turnos do taxi.';
+      }
     });
   }
